# frontend/openapi/oauth2.json
{"openapi":"3.1.0","info":{"title":"OAuth2 Authorization Server API","description":"OAuth2 Authorization Server for Compta microservices. Provides token issuance and user authentication.","license":{"name":"Apache 2.0","url":"http://springdoc.org"},"version":"1.0.0"},"servers":[{"url":"http://localhost:9000","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"UserInfo","description":"OIDC UserInfo endpoint (RFC 7662)"},{"name":"Role Management","description":"Role management endpoints"},{"name":"User Management","description":"User management endpoints"},{"name":"Client Management","description":"OAuth2 client management endpoints"},{"name":"Actuator","description":"Monitor and interact","externalDocs":{"description":"Spring Boot Actuator Web API Documentation","url":"https://docs.spring.io/spring-boot/docs/current/actuator-api/html/"}},{"name":"Token Revocation","description":"OAuth2 token revocation endpoint (RFC 7009)"},{"name":"Token Introspection","description":"OAuth2 token introspection endpoint (RFC 6819)"}],"paths":{"/api/users/{id}":{"get":{"tags":["User Management"],"summary":"Get user by ID","description":"Retrieve a specific user by ID","operationId":"getUserById","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["User Management"],"summary":"Update user","description":"Update an existing user","operationId":"updateUser","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["User Management"],"summary":"Delete user","description":"Delete a user","operationId":"deleteUser","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/roles/{id}":{"get":{"tags":["Role Management"],"summary":"Get role by ID","description":"Retrieve a specific role by ID","operationId":"getRoleById","parameters":[{"name":"id","in":"path","description":"Role ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Role Management"],"summary":"Update role","description":"Update an existing role","operationId":"updateRole","parameters":[{"name":"id","in":"path","description":"Role ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Role Management"],"summary":"Delete role","description":"Delete a role","operationId":"deleteRole","parameters":[{"name":"id","in":"path","description":"Role ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/clients/{clientId}":{"get":{"tags":["Client Management"],"summary":"Get OAuth2 client by ID","description":"Retrieve a specific OAuth2 client by its client ID","operationId":"getClientById","parameters":[{"name":"clientId","in":"path","description":"Client ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["Client Management"],"summary":"Update OAuth2 client","description":"Update an existing OAuth2 client","operationId":"updateClient","parameters":[{"name":"clientId","in":"path","description":"Client ID","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClientRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["Client Management"],"summary":"Delete OAuth2 client","description":"Delete an OAuth2 client","operationId":"deleteClient","parameters":[{"name":"clientId","in":"path","description":"Client ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/oauth2/revoke":{"post":{"tags":["Token Revocation"],"summary":"Revoke token","description":"Revoke an OAuth2 token. The token will be invalidated immediately.","operationId":"revoke","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RevocationRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/oauth2/introspect":{"post":{"tags":["Token Introspection"],"summary":"Introspect token","description":"Introspect an OAuth2 token to get its metadata and validity status","operationId":"introspect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntrospectionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/IntrospectionResponse"}}}}}}},"/api/users":{"get":{"tags":["User Management"],"summary":"Get all users","description":"Retrieve all users","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["User Management"],"summary":"Create a new user","description":"Create a new user with roles","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/roles":{"get":{"tags":["User Management"],"summary":"Get user roles","description":"Retrieve all roles for a user","operationId":"getUserRoles","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["User Management"],"summary":"Assign roles to user","description":"Assign roles to a user","operationId":"assignRoles","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"type":"string"}}}}},"required":true},"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/password":{"post":{"tags":["User Management"],"summary":"Change user password","description":"Change password for a user","operationId":"changePassword","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/users/password/reset":{"post":{"tags":["password-reset-controller"],"summary":"Initiate password reset","description":"Sends a password reset email to the user's email address","operationId":"initiatePasswordReset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}},"required":true},"responses":{"200":{"description":"Password reset email sent (even if email doesn't exist for security)","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid email format","content":{"*/*":{"schema":{"type":"string"}}}},"429":{"description":"Too many requests (rate limit exceeded)","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/password/reset/confirm":{"post":{"tags":["password-reset-controller"],"summary":"Confirm password reset","description":"Resets the user's password using a valid reset token","operationId":"confirmPasswordReset","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordResetConfirmRequest"}}},"required":true},"responses":{"200":{"description":"Password reset successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid token or password","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Token not found or expired","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/email/verify":{"post":{"tags":["email-verification-controller"],"summary":"Initiate email verification","description":"Sends a verification email to the user's email address","operationId":"initiateEmailVerification","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailVerificationRequest"}}},"required":true},"responses":{"200":{"description":"Verification email sent (even if email doesn't exist for security)","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid email format","content":{"*/*":{"schema":{"type":"string"}}}},"429":{"description":"Too many requests (rate limit exceeded)","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/users/email/verify/confirm":{"post":{"tags":["email-verification-controller"],"summary":"Confirm email verification","description":"Verifies the user's email address using a valid verification token","operationId":"confirmEmailVerification","parameters":[{"name":"token","in":"query","description":"The verification token from the email","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Email verified successfully","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Invalid or expired token","content":{"*/*":{"schema":{"type":"string"}}}},"404":{"description":"Token not found","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/roles":{"get":{"tags":["Role Management"],"summary":"Get all roles","description":"Retrieve all roles","operationId":"getAllRoles","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleWithUserCountDto"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Role Management"],"summary":"Create a new role","description":"Create a new role","operationId":"createRole","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/clients":{"get":{"tags":["Client Management"],"summary":"Get all OAuth2 clients","description":"Retrieve all registered OAuth2 clients","operationId":"getAllClients","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["Client Management"],"summary":"Create a new OAuth2 client","description":"Create a new OAuth2 registered client","operationId":"createClient","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ClientResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/clients/{clientId}/secret":{"post":{"tags":["Client Management"],"summary":"Rotate client secret","description":"Generate a new client secret for an OAuth2 client","operationId":"rotateClientSecret","parameters":[{"name":"clientId","in":"path","description":"Client ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/enable":{"patch":{"tags":["User Management"],"summary":"Enable user","description":"Enable a user account","operationId":"enableUser","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/api/users/{id}/disable":{"patch":{"tags":["User Management"],"summary":"Disable user","description":"Disable a user account","operationId":"disableUser","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/userinfo":{"get":{"tags":["UserInfo"],"summary":"Get user info","description":"Returns claims about the authenticated user","operationId":"getUserInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserInfoResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/roles/name/{name}":{"get":{"tags":["Role Management"],"summary":"Get role by name","description":"Retrieve a specific role by name","operationId":"getRoleByName","parameters":[{"name":"name","in":"path","description":"Role name","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoleResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/actuator":{"get":{"tags":["Actuator"],"summary":"Actuator root web endpoint","operationId":"links","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}},"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}}}}}}}}},"/actuator/health":{"get":{"tags":["Actuator"],"summary":"Actuator web endpoint 'health'","operationId":"health","responses":{"200":{"description":"OK","content":{"application/vnd.spring-boot.actuator.v3+json":{"schema":{"type":"object"}},"application/vnd.spring-boot.actuator.v2+json":{"schema":{"type":"object"}},"application/json":{"schema":{"type":"object"}}}}}}},"/api/users/{id}/roles/{roleId}":{"delete":{"tags":["User Management"],"summary":"Remove role from user","description":"Remove a specific role from a user","operationId":"removeRole","parameters":[{"name":"id","in":"path","description":"User ID","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"roleId","in":"path","description":"Role ID","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UpdateUserRequest":{"type":"object","properties":{"firstName":{"type":"string","maxLength":100,"minLength":0},"lastName":{"type":"string","maxLength":100,"minLength":0},"email":{"type":"string","format":"email"},"roles":{"type":"array","items":{"type":"string"}}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"enabled":{"type":"boolean"},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"},"roles":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UpdateRoleRequest":{"type":"object","properties":{"description":{"type":"string","maxLength":255,"minLength":0}}},"RoleResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"UpdateClientRequest":{"type":"object","properties":{"clientSecret":{"type":"string"},"clientName":{"type":"string","minLength":1},"clientAuthenticationMethods":{"type":"array","items":{"type":"string"}},"authorizationGrantTypes":{"type":"array","items":{"type":"string"}},"redirectUris":{"type":"array","items":{"type":"string"}},"postLogoutRedirectUris":{"type":"array","items":{"type":"string"}},"scopes":{"type":"array","items":{"type":"string"}},"requireAuthorizationConsent":{"type":"boolean"},"requireProofKey":{"type":"boolean"},"reuseRefreshTokens":{"type":"boolean"},"accessTokenTimeToLive":{"type":"string"},"authorizationCodeTimeToLive":{"type":"string"},"refreshTokenTimeToLive":{"type":"string"}},"required":["clientName"]},"ClientResponse":{"type":"object","properties":{"id":{"type":"string"},"clientId":{"type":"string"},"clientIdIssuedAt":{"type":"string","format":"date-time"},"clientName":{"type":"string"},"clientAuthenticationMethods":{"type":"array","items":{"type":"string"}},"authorizationGrantTypes":{"type":"array","items":{"type":"string"}},"redirectUris":{"type":"array","items":{"type":"string"}},"postLogoutRedirectUris":{"type":"array","items":{"type":"string"}},"scopes":{"type":"array","items":{"type":"string"}},"requireAuthorizationConsent":{"type":"boolean"},"requireProofKey":{"type":"boolean"},"reuseRefreshTokens":{"type":"boolean"},"accessTokenTimeToLiveSeconds":{"type":"integer","format":"int32"},"authorizationCodeTimeToLiveSeconds":{"type":"integer","format":"int32"},"refreshTokenTimeToLiveSeconds":{"type":"integer","format":"int32"},"clientSecretExpiresAt":{"type":"string","format":"date-time"}}},"RevocationRequest":{"type":"object","properties":{"token":{"type":"string","minLength":1},"tokenTypeHint":{"type":"string"}},"required":["token"]},"IntrospectionRequest":{"type":"object","properties":{"token":{"type":"string","minLength":1},"tokenTypeHint":{"type":"string"}},"required":["token"]},"IntrospectionResponse":{"type":"object","properties":{"active":{"type":"boolean"},"clientId":{"type":"string"},"tokenType":{"type":"string"},"exp":{"type":"string","format":"date-time"},"iat":{"type":"string","format":"date-time"},"sub":{"type":"string"},"aud":{"type":"string"},"iss":{"type":"string"},"scope":{"type":"array","items":{"type":"string"}},"username":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"CreateUserRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":50,"minLength":3},"password":{"type":"string","maxLength":100,"minLength":8},"email":{"type":"string","format":"email","minLength":1},"firstName":{"type":"string","maxLength":100,"minLength":0},"lastName":{"type":"string","maxLength":100,"minLength":0},"roles":{"type":"array","items":{"type":"string"}}},"required":["email","password","username"]},"ChangePasswordRequest":{"type":"object","properties":{"currentPassword":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":100,"minLength":8}},"required":["currentPassword","newPassword"]},"PasswordResetRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"PasswordResetConfirmRequest":{"type":"object","properties":{"token":{"type":"string","maxLength":36,"minLength":36},"password":{"type":"string","maxLength":100,"minLength":8}},"required":["password","token"]},"EmailVerificationRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"CreateRoleRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":2},"description":{"type":"string","maxLength":255,"minLength":0}},"required":["name"]},"CreateClientRequest":{"type":"object","properties":{"clientId":{"type":"string","minLength":1},"clientSecret":{"type":"string"},"clientName":{"type":"string","minLength":1},"clientAuthenticationMethods":{"type":"array","items":{"type":"string"}},"authorizationGrantTypes":{"type":"array","items":{"type":"string"}},"redirectUris":{"type":"array","items":{"type":"string"}},"postLogoutRedirectUris":{"type":"array","items":{"type":"string"}},"scopes":{"type":"array","items":{"type":"string"}},"requireAuthorizationConsent":{"type":"boolean"},"requireProofKey":{"type":"boolean"},"reuseRefreshTokens":{"type":"boolean"},"accessTokenTimeToLive":{"type":"string"},"authorizationCodeTimeToLive":{"type":"string"},"refreshTokenTimeToLive":{"type":"string"}},"required":["authorizationGrantTypes","clientAuthenticationMethods","clientId","clientName","redirectUris","scopes"]},"UserInfoResponse":{"type":"object","properties":{"sub":{"type":"string"},"name":{"type":"string"},"givenName":{"type":"string"},"familyName":{"type":"string"},"email":{"type":"string"},"emailVerified":{"type":"boolean"},"roles":{"type":"array","items":{"type":"string"}},"tenantId":{"type":"string"}}},"RoleWithUserCountDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"userCount":{"type":"integer","format":"int32"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT token for authentication","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}